+++ apps/shared/toolbox_helpers.cpp
@@ -5,7 +5,7 @@ namespace Shared {
 namespace ToolboxHelpers {
 
 int CursorIndexInCommand(const char * text) {
-  for (int i = 0; i < strlen(text); i++) {
+  for (size_t i = 0; i < strlen(text); i++) {
     if (text[i] == '(') {
       return i + 1;
     } else if (text[i] == '.') {
@@ -19,7 +19,7 @@ void TextToInsertForCommandMessage(I18n::Message message, char * buffer) {
   const char * messageText = I18n::translate(message);
   int currentNewTextIndex = 0;
   int numberOfOpenBrackets = 0;
-  for (int i = 0; i < strlen(messageText); i++) {
+  for (size_t i = 0; i < strlen(messageText); i++) {
     if (messageText[i] == '.') {
       currentNewTextIndex = 0;
       numberOfOpenBrackets = 0;
